Refactor Saloon\Contracts\Validators to Saloon\Contracts\Testing
