+++ src/Contracts/Testing/ValidatorCheck.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Saloon\Contracts\Validators;
+namespace Saloon\Contracts\Testing;
 
 interface ValidatorCheck
 {
